backend/requirements.txt
# Core FastAPI
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.6.1

# File uploads and forms
python-multipart==0.0.19

# Environment and configuration
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
asyncpg==0.30.0

# Supabase integration
supabase==2.10.0
postgrest==0.18.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jwt==4.1.0

# HTTP client
httpx==0.28.1
aiohttp==3.11.10

# Redis and caching
redis==5.2.1
aioredis==2.0.1

# Task queue
celery==5.4.0

# Monitoring and logging
structlog==24.5.0
prometheus-client==0.21.1

# Rate limiting
slowapi==0.1.9
limits==3.14.0

# CORS
python-cors==1.7.0

# Data validation and serialization
email-validator==2.2.0
phonenumbers==8.13.51

# Date and time handling
python-dateutil==2.9.0.post0

# GitHub API
PyGithub==2.5.0

# OpenAI integration
openai==1.58.1

# Development and testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
mypy==1.14.0
